+++ Service/Sources/Base/DataSource/BaseRemoteDataSource.swift
@@ -132,12 +132,3 @@ private extension _Concurrency.Task where Failure == Error {
         }
     }
 }
-
-extension String {
-    func toBitgouelDate() -> Date {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss"
-        formatter.locale = Locale(identifier: "ko_kr")
-        return formatter.date(from: self) ?? .init()
-    }
-}
