+++ Service/Sources/Utils/Date/DateUtil.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+public extension String {
+    func toBitgouelDate() -> Date {
+        let formatter = DateFormatter()
+        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss"
+        formatter.locale = Locale(identifier: "ko_kr")
+        return formatter.date(from: self) ?? .init()
+    }
+}
